+++ docs/source/dev_documentation/changelog.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 3.0.1 - 2025-04-15
+
+### Fixed
+
+- Ledger account instantiation not detected
+- Missing ledger dependencies
+- Missing hash signing option when using ledger account
+
+## Added
+
+- Documentation page for the chain-simulator
+
 ## 3.0.0 - 2025-04-04
 
 ### Fixed